fix(x11): widen XID to u64 in event routing for 32-bit targets
self.window is an X11 Window (c_ulong), which is u32 on 32-bit targets
(i686/armv7). Comparing it directly against `target: u64` (the cast
XAnyEvent.window) failed to compile there with "can't compare u64 with
u32". Cast self.window to u64 in the comparison -- a no-op on 64-bit,
a widening on 32-bit. Unblocks the 32-bit cross-builds in CI.

+++ dll/src/desktop/shell2/linux/x11/mod.rs
@@ -733,7 +733,9 @@ impl X11Window {
                 unsafe { (self.xlib.XNextEvent)(self.display, &mut event) };
 
                 let target = unsafe { event.any.window } as u64;
-                if target == self.window {
+                // self.window is an XID (c_ulong) = u32 on 32-bit targets, so
+                // widen it to match `target: u64` (no-op on 64-bit).
+                if target == self.window as u64 {
                     self.handle_event(&mut event);
                 } else if let Some(wptr) = unsafe { super::registry::get_window(target) } {
                     // Dispatch to the child window that owns this X window.
